backend/pyproject.toml
[project]
name = "claim-agent"
version = "0.1.0"
description = "Insurance claim preparation agent — helps decide whether to file and preps for the call"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = ["integration: tests requiring AWS Bedrock"]
addopts = "-m 'not integration'"
